requirements.txt
Pillow == 9.1.1
